refactor(types): rename `_typehints` to `_typing`
