+++ pyrogram/methods/account/add_profile_audio.py
@@ -25,7 +25,7 @@ from collections.abc import Callable
 
 import pyrogram
 from pyrogram import StopTransmission, raw, utils
-from pyrogram._typehints import PathType
+from pyrogram._typing import PathType
 from pyrogram.errors import FilePartMissing
 from pyrogram.file_id import FileType
