+++ pyrogram/methods/messages/download_media.py
@@ -25,7 +25,7 @@ from typing import TYPE_CHECKING, BinaryIO, Literal, overload
 
 import pyrogram
 from pyrogram import types, utils
-from pyrogram._typehints import PathType
+from pyrogram._typing import PathType
 from pyrogram.file_id import PHOTO_TYPES, FileId, FileType
 
 if TYPE_CHECKING:
