+++ pyrogram/methods/messages/send_live_photo.py
@@ -25,7 +25,7 @@ from collections.abc import Callable
 
 import pyrogram
 from pyrogram import StopTransmission, enums, raw, types, utils
-from pyrogram._typehints import PathType
+from pyrogram._typing import PathType
 from pyrogram.errors import FilePartMissing
 
 log = logging.getLogger(__name__)
