+++ pyrogram/types/input_content/input_chat_photo.py
@@ -22,7 +22,7 @@ from typing import BinaryIO, cast
 
 import pyrogram
 from pyrogram import raw, utils
-from pyrogram._typehints import PathType
+from pyrogram._typing import PathType
 from pyrogram.file_id import FileType
 
 from ..object import Object
