+++ pyrogram/types/messages_and_media/message.py
@@ -28,7 +28,7 @@ from collections.abc import Callable
 
 import pyrogram
 from pyrogram import enums, raw, types, utils
-from pyrogram._typehints import PathType
+from pyrogram._typing import PathType
 from pyrogram.errors import (
     ChannelForumMissing,
     ChannelInvalid,
